+++ paynlsdk2/api/refund/info.py
@@ -1,57 +1,69 @@
 import json
 
-from marshmallow import Schema, fields, post_load
+from marshmallow import Schema, fields, post_load, pre_load
 
-from paynlsdk.api.requestbase import RequestBase
-from paynlsdk.api.responsebase import ResponseBase
-from paynlsdk.objects import ErrorSchema
-from paynlsdk.validators import ParamValidator
+from paynlsdk2.api.requestbase import RequestBase
+from paynlsdk2.api.responsebase import ResponseBase
+from paynlsdk2.objects import ErrorSchema, RefundInfo, RefundInfoSchema
+from paynlsdk2.validators import ParamValidator
 
 
 class Response(ResponseBase):
+    # type: (ResponseBase)
     """
-    Response object for the Transaction::decline API
+    Response object for the Refund::info API
 
-    :param bool result: Result of the API call
-    :param str message: message related to the API call
+    :param str refund_id: Refund ID
+    :param RefundInfo refund: Refund information
     """
     def __init__(self,
-                 result=None,
-                 message=None,
+                 refund_id=None,
+                 refund=None,
                  *args, **kwargs):
-        # type: (bool, str) -> None
-        self.result = result
-        self.message = message
+        # type: (str, RefundInfo *str, **Any) -> None
+        self.refund_id = refund_id
+        self.refund = refund
         super(Response, self).__init__(**kwargs)
 
     def __repr__(self):
-        # type: () -> str
-        return self.__dict__.__str__()
+        return str(self.__dict__)
+
+    def is_refunded(self):
+        # type: () -> bool
+        """
+        Check if refund is processed
+        :return: indication whether the refund has been processed or not
+        :rtype: bool
+        """
+        return self.refund.status_name == 'Verwerkt'
 
 
 class ResponseSchema(Schema):
-    request = fields.Nested(ErrorSchema)
-    message = fields.String(required=False)
+    request = fields.Nested(ErrorSchema, required=True)
+    refund_id = fields.String(required=True, load_from='refundId')
+    refund = fields.Nested(RefundInfoSchema, required=True, load_from='refund', allow_none=True)
+
+    @pre_load
+    def pre_processor(self, data):
+        if data['refund'] == '':
+            data['refund'] = None
+        return data
 
     @post_load
     def create_response(self, data):
         # type: (dict) -> Response
-        #  We will map the result to a Response internal value
-        data['result'] = data['request'].result
         return Response(**data)
 
 
 class Request(RequestBase):
     """
-    Request object for the Transaction::decline API
+    Request object for the Refund::info API
 
-    :param str order_id: transaction ID (wrongly names order_id)
-    :param str entrance_code: Entrance code
+    :param str refund_id: Refund ID
     """
-    def __init__(self, order_id=None, entrance_code=None):
-        # type: (str, str) -> None
-        self.order_id = order_id
-        self.entrance_code = entrance_code
+    def __init__(self, refund_id=None):
+        # type: (str) -> None
+        self.refund_id = refund_id
         super(Request, self).__init__()
 
     def requires_api_token(self):
@@ -64,15 +76,15 @@ class Request(RequestBase):
 
     def get_version(self):
         # type: () -> int
-        return 12
+        return 3
 
     def get_controller(self):
         # type: () -> str
-        return 'Transaction'
+        return 'Refund'
 
     def get_method(self):
         # type: () -> str
-        return 'decline'
+        return 'info'
 
     def get_query_string(self):
         # type: () -> str
@@ -81,17 +93,16 @@ class Request(RequestBase):
     def get_parameters(self):
         # type: () -> dict
         # Validation
-        ParamValidator.assert_not_empty(self.order_id, 'order_id')
+        ParamValidator.assert_not_empty(self.refund_id, 'refund_id')
         # Get default api parameters
         rs = self.get_std_parameters()
         # Add own parameters
-        rs['orderId'] = self.order_id
-        if ParamValidator.not_empty(self.entrance_code):
-            rs['entranceCode'] = self.entrance_code
+        rs['refundId'] = self.refund_id
         return rs
 
     @RequestBase.raw_response.setter
     def raw_response(self, raw_response):
+        # type: (str) -> None
         self._raw_response = raw_response
         # Do error checking.
         rs = json.loads(self.raw_response)
@@ -107,12 +118,13 @@ class Request(RequestBase):
         Return the API :class:`Response` for the validation request
 
         :return: The API response
-        :rtype: paynlsdk.api.transaction.decline.Response
+        :rtype: paynlsdk2.api.refund.info.Response
         """
         return self._response
 
     @response.setter
     def response(self, response):
+        # type: (Response) -> None
         # print('{}::respone.setter'.format(self.__module__ + '.' + self.__class__.__qualname__))
         self._response = response
