+++ paynlsdk2/api/transaction/approve.py
@@ -2,10 +2,10 @@ import json
 
 from marshmallow import Schema, fields, post_load
 
-from paynlsdk.api.requestbase import RequestBase
-from paynlsdk.api.responsebase import ResponseBase
-from paynlsdk.objects import ErrorSchema
-from paynlsdk.validators import ParamValidator
+from paynlsdk2.api.requestbase import RequestBase
+from paynlsdk2.api.responsebase import ResponseBase
+from paynlsdk2.objects import ErrorSchema
+from paynlsdk2.validators import ParamValidator
 
 
 class Response(ResponseBase):
@@ -107,7 +107,7 @@ class Request(RequestBase):
         Return the API :class:`Response` for the validation request
 
         :return: The API response
-        :rtype: paynlsdk.api.transaction.approve.Response
+        :rtype: paynlsdk2.api.transaction.approve.Response
         """
         return self._response
