+++ paynlsdk2/client/refund.py
@@ -1,19 +1,19 @@
-from paynlsdk.api.client import APIClient
+from paynlsdk2.api.client import APIClient
 
 
 class Refund(object):
     @staticmethod
     def info(refund_id):
-        # type: (str) -> paynlsdk.api.refund.info.Response
+        # type: (str) -> paynlsdk2.api.refund.info.Response
         """
         Return refund info
 
         :param refund_id: Refund ID (starts wih "RF-")
         :type refund_id: str
         :return: Info Response
-        :rtype: paynlsdk.api.refund.info.Response
+        :rtype: paynlsdk2.api.refund.info.Response
         """
-        from paynlsdk.api.refund.info import Request
+        from paynlsdk2.api.refund.info import Request
         client = APIClient()
         request = Request(refund_id)
         client.perform_request(request)
@@ -27,7 +27,7 @@ class Refund(object):
                     products={},
                     vat_percentage=None,
                     exchange_url=None):
-        # type: (str, int, str, str, dict, float, str) -> paynlsdk.api.refund.transaction.Response
+        # type: (str, int, str, str, dict, float, str) -> paynlsdk2.api.refund.transaction.Response
         """
         Refund a transaction
 
@@ -47,9 +47,9 @@ class Refund(object):
         :param exchange_url: URL for the exchange call
         :type exchange_url: str
         :return: Transaction refund response
-        :rtype: paynlsdk.api.refund.transaction.Response
+        :rtype: paynlsdk2.api.refund.transaction.Response
         """
-        from paynlsdk.api.refund.transaction import Request
+        from paynlsdk2.api.refund.transaction import Request
         client = APIClient()
         request = Request(transaction_id, amount, description, process_date, products, vat_percentage, exchange_url)
         client.perform_request(request)
@@ -57,24 +57,24 @@ class Refund(object):
 
     @staticmethod
     def info_request():
-        # type: () -> paynlsdk.api.refund.info.Request
+        # type: () -> paynlsdk2.api.refund.info.Request
         """
         Get a refund info request instance
 
         :return: The request object that can be configured
-        :rtype: paynlsdk.api.refund.info.Request
+        :rtype: paynlsdk2.api.refund.info.Request
         """
-        from paynlsdk.api.refund.info import Request
+        from paynlsdk2.api.refund.info import Request
         return Request()
 
     @staticmethod
     def transaction_request():
-        # type: () -> paynlsdk.api.refund.info.Request
+        # type: () -> paynlsdk2.api.refund.info.Request
         """
         Get a refund transaction request instance
 
         :return: The request object that can be configured
-        :rtype: paynlsdk.api.refund.transaction.Request
+        :rtype: paynlsdk2.api.refund.transaction.Request
         """
-        from paynlsdk.api.transaction.info import Request
+        from paynlsdk2.api.transaction.info import Request
         return Request()
