+++ paynlsdk2/client/validate.py
@@ -1,4 +1,4 @@
-from paynlsdk.api.client import APIClient
+from paynlsdk2.api.client import APIClient
 
 
 class Validate(object):
@@ -11,36 +11,36 @@ class Validate(object):
         :param ip_address: IP address
         :type ip_address: str
         :return: Request instance
-        :rtype: paynlsdk.api.validate.payserverip.Request
+        :rtype: paynlsdk2.api.validate.payserverip.Request
         """
         response = Validate.pay_server_ip_response(ip_address)
         return response.result
 
     @staticmethod
     def pay_server_ip_request():
-        # type: () -> paynlsdk.api.validate.payserverip.Request
+        # type: () -> paynlsdk2.api.validate.payserverip.Request
         """
-        Get a Pay server IP validation :class:`paynlsdk.api.validate.payserverip.Request` instance
+        Get a Pay server IP validation :class:`paynlsdk2.api.validate.payserverip.Request` instance
 
         :return: Request instance
-        :rtype: paynlsdk.api.validate.payserverip.Request
+        :rtype: paynlsdk2.api.validate.payserverip.Request
         """
-        from paynlsdk.api.validate.payserverip import Request
+        from paynlsdk2.api.validate.payserverip import Request
         request = Request()
         return request
 
     @staticmethod
     def pay_server_ip_response(ip_address):
-        # type: (str) -> paynlsdk.api.validate.payserverip.Response
+        # type: (str) -> paynlsdk2.api.validate.payserverip.Response
         """
-        Get a Pay server IP validation :class:`paynlsdk.api.validate.payserverip.Response` instance
+        Get a Pay server IP validation :class:`paynlsdk2.api.validate.payserverip.Response` instance
 
         :param ip_address: IP address
         :type ip_address: str
         :return: Response instance
-        :rtype: paynlsdk.api.validate.payserverip.Response
+        :rtype: paynlsdk2.api.validate.payserverip.Response
         """
-        from paynlsdk.api.validate.payserverip import Request
+        from paynlsdk2.api.validate.payserverip import Request
         client = APIClient()
         request = Request(ip_address)
         client.perform_request(request)
